Removing a bunch of "dangerous" commands
Newset sonnet likes to write test scripts, and with rm as a dangerous
command, it cannot clean up after itself. mv prevents us from a lot of
simple refactors/renames, and the redirects allow us to write or append
to files quickly in a way the model is trained to do. might as well
allow it, makes less of a mess.

the broader issue is that we're not parsing these commands or
intercepting syscalls, so they allow for egress from the filesystem
sandbox (supposed to be the tree starting at cwd). This merits more
work, but biasing towards more functional loops.